update all events to macros
