+++ gasometer/src/lib.rs
@@ -5,11 +5,26 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
+#[cfg(feature = "tracing")]
+pub mod tracing;
+
+#[cfg(feature = "tracing")]
+macro_rules! event {
+	($x:expr) => {
+		use crate::tracing::Event::*;
+		$x.emit();
+	}
+}
+
+#[cfg(not(feature = "tracing"))]
+macro_rules! event {
+	($x:expr) => { }
+}
+
 mod consts;
 mod costs;
 mod memory;
 mod utils;
-pub mod tracing;
 
 use core::cmp::max;
 use primitive_types::{H160, H256, U256};
@@ -26,6 +41,14 @@ macro_rules! try_or_fail {
 			},
 		}
 	)
+}
+
+#[derive(Debug, Copy, Clone)]
+pub struct Snapshot {
+    pub gas_limit: u64,
+    pub memory_gas: u64,
+	pub used_gas: u64,
+	pub refunded_gas: i64,
 }
 
 /// EVM gasometer.
@@ -116,10 +139,10 @@ impl<'config> Gasometer<'config> {
 		&mut self,
 		cost: u64,
 	) -> Result<(), ExitError> {
-		tracing::Event::RecordCost {
+		event!(RecordCost {
 			cost,
 			snapshot: self.snapshot()?,
-		}.emit();
+		});
 
 		let all_gas_cost = self.total_used_gas() + cost;
 		if self.gas_limit < all_gas_cost {
@@ -137,10 +160,10 @@ impl<'config> Gasometer<'config> {
 		&mut self,
 		refund: i64,
 	) -> Result<(), ExitError> {
-		tracing::Event::RecordRefund {
+		event!(RecordRefund {
 			refund,			
 			snapshot: self.snapshot()?,
-		}.emit();
+		});
 
 		self.inner_mut()?.refunded_gas += refund;
 		Ok(())
@@ -172,12 +195,12 @@ impl<'config> Gasometer<'config> {
 		let gas_refund = self.inner_mut()?.gas_refund(cost);
 		let used_gas = self.inner_mut()?.used_gas;
 
-		tracing::Event::RecordDynamicCost {
+		event!(RecordDynamicCost {
 			gas_cost,
 			memory_gas,
 			gas_refund, 
 			snapshot: self.snapshot()?,
-		}.emit();
+		});
 
 		let all_gas_cost = memory_gas + used_gas + gas_cost;
 		if self.gas_limit < all_gas_cost {
@@ -201,10 +224,10 @@ impl<'config> Gasometer<'config> {
 		&mut self,
 		stipend: u64,
 	) -> Result<(), ExitError> {
-		tracing::Event::RecordStipend {
+		event!(RecordStipend {
 			stipend,
 			snapshot: self.snapshot()?,
-		}.emit();
+		});
 
 		self.inner_mut()?.used_gas -= stipend;
 		Ok(())
@@ -228,10 +251,10 @@ impl<'config> Gasometer<'config> {
 			},
 		};
 
-		tracing::Event::RecordTransaction {
+		event!(RecordTransaction {
 			cost: gas_cost,
 			snapshot: self.snapshot()?,
-		}.emit();
+		});
 
 		if self.gas() < gas_cost {
 			self.inner = Err(ExitError::OutOfGas);
@@ -242,9 +265,9 @@ impl<'config> Gasometer<'config> {
 		Ok(())
 	}
 
-	fn snapshot(&self) -> Result<tracing::Snapshot, ExitError> {
+	pub fn snapshot(&self) -> Result<Snapshot, ExitError> {
 		let inner = self.inner.as_ref().map_err(|e| e.clone())?;
-		Ok(tracing::Snapshot {
+		Ok(Snapshot {
 			gas_limit: self.gas_limit,
 			memory_gas: inner.memory_gas,
 			used_gas: inner.used_gas,
