+++ src/executor/stack/mod.rs
@@ -7,7 +7,7 @@ use alloc::{rc::Rc, vec::Vec};
 use primitive_types::{U256, H256, H160};
 use sha3::{Keccak256, Digest};
 use crate::{ExitError, Stack, Opcode, Capture, Handler, Transfer,
-			Context, CreateScheme, Runtime, ExitReason, ExitSucceed, Config, tracing};
+			Context, CreateScheme, Runtime, ExitReason, ExitSucceed, Config};
 use ethereum::Log;
 use crate::gasometer::{self, Gasometer};
 
@@ -338,14 +338,14 @@ impl<'config, S: StackState<'config>> StackExecutor<'config, S> {
 
 		let address = self.create_address(scheme);
 
-		tracing::Event::Create {
+		event!(Create {
 			caller,
 			address,
 			scheme,
 			value,
 			init_code: &init_code,
 			target_gas
-		}.emit();
+		});
 
 		if let Some(depth) = self.state.metadata().depth {
 			if depth > self.config.call_stack_limit {
@@ -493,14 +493,14 @@ impl<'config, S: StackState<'config>> StackExecutor<'config, S> {
 			gas - gas / 64
 		}
 
-		tracing::Event::Call {
+		event!(Call {
 			code_address,
 			transfer: &transfer,
 			input: &input,
 			target_gas,
 			is_static,
 			context: &context,
-		}.emit();
+		});
 
 		let after_gas = if take_l64 && self.config.call_l64_after_gas {
 			if self.config.estimate {
@@ -678,11 +678,11 @@ impl<'config, S: StackState<'config>> Handler for StackExecutor<'config, S> {
 	fn mark_delete(&mut self, address: H160, target: H160) -> Result<(), ExitError> {
 		let balance = self.balance(address);
 
-		tracing::Event::Suicide {
+		event!(Suicide {
 			target,
 			address,
 			balance,
-		}.emit();
+		});
 
 		self.state.transfer(Transfer {
 			source: address,
